+++ src/lib/file/FileItemTable.svelte
@@ -4,6 +4,7 @@ import { derived } from 'svelte/store'
 import { annotatedFileData } from './annotated.js'
 import { fileItems, snippet_target } from './file.js';
 import Channel from '../update/Channel.svelte';
+import Time from '../update/Time.svelte';
 
 export let img_src;
 export let selected;
@@ -148,46 +149,6 @@ async function closeOut() {
         .then(response => response.json())
         .then(data => annotatedFileData.set(data))
     sortedFileItems = sortFileItems();
-}
-
-let time_editable = -1;
-let new_time = '';
-let old_time = '';
-function TimeEditable(time, file_date, index) {
-    return () => {
-        time_editable = index;
-        if (time != null) {
-            new_time = time;
-            old_time = time;
-        } else {
-            new_time = file_date;
-            old_time = time;
-        }
-    }
-}
-async function updateTime(box_id) {
-    if (new_time != old_time) {
-        let date = new Date(new_time);
-        date = new Date(date.toLocaleString('en-US', {timeZone: 'Greenwich'}));
-        await fetch('http://localhost:5000/time/update?api_key=' + import.meta.env.VITE_API_KEY, {
-            method: 'POST',
-            body: JSON.stringify({
-                'id': box_id,
-                'year': date.getFullYear(),
-                'month': date.getMonth() + 1,
-                'day': date.getDate(),
-                'hour': date.getHours(),
-                'minute': date.getMinutes()
-            })
-        })
-        await fetch('http://localhost:5000/file/?file_id=' + selected + '&api_key=' + import.meta.env.VITE_API_KEY)
-            .then(response => response.json())
-            .then(data => annotatedFileData.set(data))
-        sortedFileItems = sortFileItems();
-    }
-    time_editable = -1;
-    new_time = '';
-    old_time = '';
 }
 
 let duration_editable = -1;
@@ -359,14 +320,8 @@ async function updateVCRCode(box_id, file_date) {
             <td>
                 <Channel closeOut={closeOut} item={fileItem} index={index} show_title={false}/>
             </td>
-            <td on:dblclick={TimeEditable(fileItem.time, fileItem.file_date, index)}>
-                {#if index != time_editable}
-                {fileItem.time}
-                {:else}
-                <form on:submit|preventDefault={(e) => updateTime(fileItem.box_id)}>
-                    <input id="time_update" bind:value={new_time} />
-                </form>
-                {/if}
+            <td>
+                <Time closeOut={closeOut} item={fileItem} index={index} show_title={false}/>
             </td>
             <td on:dblclick={DurationEditable(fileItem.duration_minutes, index)}>
                 {#if index != duration_editable}
